requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
transformers>=4.30.0
scikit-learn>=1.3.0
numpy>=1.24.0
torch>=2.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0
sentence-transformers>=2.2.0
plotly>=5.17.0
kaleido>=0.2.1
reportlab>=4.0.0

